+++ src/core/server/athena/main.ts
@@ -16,8 +16,6 @@ export const DEFAULT_CONFIG = {
     MAX_INTERACTION_DISTANCE: 3, // The maximum distance the player can stand away from a interaction object.
     // Vehicle Properties
     TIME_BETWEEN_VEHICLE_SAVES: 30000, // 30s
-    SPAWN_ALL_VEHICLES_ON_START: true, // Will spawn all vehicles unless they're in a garage.
-    SPAWN_VEHICLES_ON_JOIN: false, // Will spawn all player vehicles not in a garage on join. Does not work if SPAWN_ALL_VEHICLES_ON_START is enabled.
     DESPAWN_VEHICLES_ON_LOGOUT: false, // Will despawn all player vehicles on logout.
     VEHICLE_SPAWN_TIMEOUT: 24, // Hours that must pass since a vehicle was last used to avoid spawning on server restart.
     VEHICLE_MAX_DISTANCE_TO_ENTER: SHARED_CONFIG.MAX_VEHICLE_INTERACTION_RANGE, // Max distance to enter a vehicle.