+++ src/core/server/vehicle/ownership.ts
@@ -1,5 +1,6 @@
 import * as alt from 'alt-server';
 import * as Athena from '@AthenaServer/api';
+import Database from '@stuyk/ezmongodb';
 
 /**
  * A catch all handler for vehicle ownership.
@@ -134,4 +135,97 @@ export function getAsPlayer(vehicle: alt.Vehicle): alt.Player | undefined {
     }
 
     return Athena.getters.player.byDatabaseID(data.owner);
+}
+
+/**
+ * Add a character to the owned vehicle
+ *
+ * @export
+ * @param {alt.Vehicle} vehicle
+ * @param {alt.Player} player
+ * @returns {Promise<boolean>}
+ */
+export async function addCharacter(vehicle: alt.Vehicle, player: alt.Player): Promise<boolean>;
+/**
+ * Add a character to the owned vehicle
+ *
+ * @export
+ * @param {alt.Vehicle} vehicle
+ * @param {string} id
+ * @returns {Promise<boolean>}
+ */
+export async function addCharacter(vehicle: alt.Vehicle, id: string): Promise<boolean>;
+/**
+ * Add a character or player to the owned vehicle
+ *
+ * @export
+ * @param {alt.Vehicle} vehicle
+ * @param {(alt.Player | string)} playerOrId
+ * @returns {Promise<boolean>}
+ */
+export async function addCharacter(vehicle: alt.Vehicle, playerOrId: alt.Player | string): Promise<boolean> {
+    const data = Athena.document.vehicle.get(vehicle);
+    if (typeof data === 'undefined') {
+        return false;
+    }
+
+    let _id: string;
+    if (playerOrId instanceof alt.Player) {
+        const playerData = Athena.document.character.get(playerOrId);
+        if (typeof playerData === 'undefined') {
+            return false;
+        }
+
+        _id = playerData._id.toString();
+    }
+
+    if (typeof playerOrId === 'string') {
+        _id = playerOrId;
+    }
+
+    if (!data.keys) {
+        data.keys = [];
+    }
+
+    const index = data.keys.findIndex((x) => x === _id);
+    if (index <= -1) {
+        data.keys.push(_id);
+    }
+
+    return await Database.updatePartialData(
+        data._id.toString(),
+        { keys: data.keys },
+        Athena.database.collections.Vehicles,
+    );
+}
+
+/**
+ * Remove a character from the owned vehicle keys.
+ *
+ * @export
+ * @param {alt.Vehicle} vehicle
+ * @param {(alt.Player | string)} playerOrId
+ * @returns {Promise<boolean>}
+ */
+export async function removeCharacter(vehicle: alt.Vehicle, _id: string): Promise<boolean> {
+    const data = Athena.document.vehicle.get(vehicle);
+    if (typeof data === 'undefined') {
+        return false;
+    }
+
+    if (!data.keys) {
+        data.keys = [];
+    }
+
+    const index = data.keys.findIndex((x) => x === _id);
+    if (index <= -1) {
+        return false;
+    }
+
+    data.keys.splice(index, 1);
+    return await Database.updatePartialData(
+        data._id.toString(),
+        { keys: data.keys },
+        Athena.database.collections.Vehicles,
+    );
 }